ナビゲーション仮実装,VECTOR3クラス強化,initialize.txtの強化
ナビゲーションは未テスト
VECTOR3に各種演算子を実装
initialize.txtの書式を変更
　　　　一行目：使用するタスクを列挙
　　　　二行目以降：最初に実行するコマンドを列挙

+++ motor.h
@@ -114,7 +114,7 @@ public:
 		//PID制御用パラメータを設定
 		void set(double p,double i,double d);
 		//PID制御を開始する(ジャイロのZ回転軸の角度が一定になるように制御する)
-		void startPID(int angle,int power);
+		void startPID(double angle,int power);
 
         MotorDrive();
         ~MotorDrive();
